refactor(ui): move show_buy_skill_dlg / show_buy_emph_dlg to skills tab
These two BuyAdvDialog launchers were placed in AdvanceMixin together
with the other show_*/learn_* methods, but unlike the rest they are
NOT triggered by a rank-up nicebar check -- they are wired directly to
the Tab 2 skills toolbar (vtb.addButton in build_ui_page_2). Move them
to SkillsTabMixin where the call sites live.

The remaining show_/learn_ methods in AdvanceMixin (show_advance_rank_dlg,
show_select_affinity, show_select_deficiency, learn_next_school_spells,
learn_next_free_kiho, act_choose_skills) all stay -- they form the
"check_X -> nicebar -> show_Y/learn_Y" pair pattern with their
matching check_ methods.
