+++ l5r/ui/tabs/skills.py
@@ -11,6 +11,7 @@
 
 from qtpy import QtCore, QtGui
 
+import l5r.dialogs as dialogs
 import l5r.models as models
 import l5r.widgets as widgets
 
@@ -92,3 +93,20 @@ class SkillsTabMixin:
             if idx.isValid():
                 sm_.setCurrentIndex(idx, (QtCore.QItemSelectionModel.Select |
                                           QtCore.QItemSelectionModel.Rows))
+
+    def show_buy_skill_dlg(self):
+        dlg = dialogs.BuyAdvDialog(self.pc, 'skill', self)
+        dlg.exec_()
+        self.update_from_model()
+
+    def show_buy_emph_dlg(self):
+        # get selected skill
+        sm_ = self.skill_table_view.selectionModel()
+        if sm_.hasSelection():
+            model_ = self.skill_table_view.model()
+            skill_id = model_.data(sm_.currentIndex(), QtCore.Qt.UserRole)
+
+            dlg = dialogs.BuyAdvDialog(self.pc, 'emph', self)
+            dlg.fix_skill_id(skill_id)
+            dlg.exec_()
+            self.update_from_model()
